ml-models/requirements-global.txt
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git
open-clip-torch>=2.20.0
transformers>=4.30.0
easyocr>=1.7.0
faiss-cpu>=1.7.4
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
